+++ NRDB/DeckListViewController.m
@@ -854,8 +854,8 @@ enum { NAME_ALERT = 1, SWITCH_ALERT };
         cell.copiesLabel.textColor = [UIColor blackColor];
         if ([card.setCode isEqualToString:@"core"])
         {
-            int cores = [[NSUserDefaults standardUserDefaults] integerForKey:NUM_CORES];
-            int owned = cores * card.quantity;
+            NSInteger cores = [[NSUserDefaults standardUserDefaults] integerForKey:NUM_CORES];
+            NSInteger owned = cores * card.quantity;
             
             if (owned < cc.count)
             {
